+++ packages/article-main-comment/src/article-main-comment.js
@@ -30,9 +30,11 @@ class ArticlePage extends Component {
       standfirst
     } = article;
 
+    const authorImage = author && author.image ? author.image : null;
+
     return (
       <ArticleHeader
-        authorImage={author.image}
+        authorImage={authorImage}
         byline={byline}
         flags={flags}
         headline={getHeadline(headline, shortHeadline)}