+++ packages/article-main-comment/src/styles/responsive.web.js
@@ -90,7 +90,7 @@ export const Seperator = styled(View)`
   display: none;
 
   @media (min-width: ${breakpoints.medium}px) {
-    background-color: ${colours.functional.keyline}
+    background-color: ${colours.functional.keyline};
     display: flex;
     height: ${spacing(3)};
     margin: 0 ${spacing(2)};