+++ packages/caption/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.2.1"></a>
+## [2.2.1](https://github.com/newsuk/times-components/compare/@times-components/caption@2.2.0...@times-components/caption@2.2.1) (2018-11-27)
+
+**Note:** Version bump only for package @times-components/caption
+
+
+
+
+
 <a name="2.2.0"></a>
 # [2.2.0](https://github.com/newsuk/times-components/compare/@times-components/caption@2.1.52...@times-components/caption@2.2.0) (2018-11-26)