+++ packages/link/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.1.12"></a>
+## [3.1.12](https://github.com/newsuk/times-components/compare/@times-components/link@3.1.11...@times-components/link@3.1.12) (2018-11-27)
+
+**Note:** Version bump only for package @times-components/link
+
+
+
+
+
 <a name="3.1.11"></a>
 ## [3.1.11](https://github.com/newsuk/times-components/compare/@times-components/link@3.1.10...@times-components/link@3.1.11) (2018-11-26)