+++ packages/provider-queries/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.8.5"></a>
+## [1.8.5](https://github.com/newsuk/times-components/compare/@times-components/provider-queries@1.8.4...@times-components/provider-queries@1.8.5) (2018-11-27)
+
+**Note:** Version bump only for package @times-components/provider-queries
+
+
+
+
+
 <a name="1.8.4"></a>
 ## [1.8.4](https://github.com/newsuk/times-components/compare/@times-components/provider-queries@1.8.3...@times-components/provider-queries@1.8.4) (2018-11-21)