+++ packages/pull-quote/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.2.47"></a>
+## [3.2.47](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@3.2.46...@times-components/pull-quote@3.2.47) (2018-11-27)
+
+**Note:** Version bump only for package @times-components/pull-quote
+
+
+
+
+
 <a name="3.2.46"></a>
 ## [3.2.46](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@3.2.45...@times-components/pull-quote@3.2.46) (2018-11-26)