+++ packages/video/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.1.21"></a>
+## [4.1.21](https://github.com/newsuk/times-components/compare/@times-components/video@4.1.20...@times-components/video@4.1.21) (2018-11-27)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
 <a name="4.1.20"></a>
 ## [4.1.20](https://github.com/newsuk/times-components/compare/@times-components/video@4.1.19...@times-components/video@4.1.20) (2018-11-26)